pyproject.toml
[project]
name = "asisto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "firebase-admin>=7.2.0",
    "google-adk[vertexai]>=1.27.1",
    "pyyaml>=6.0.3",
    "uvicorn>=0.41.0",
]
